move client code into sdk package
