+++ cmd/bank-vaults/init.go
@@ -15,7 +15,7 @@
 package main
 
 import (
-	"github.com/banzaicloud/bank-vaults/pkg/vault"
+	"github.com/banzaicloud/bank-vaults/pkg/sdk/vault"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
