+++ cmd/bank-vaults/unseal.go
@@ -18,7 +18,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/banzaicloud/bank-vaults/pkg/vault"
+	"github.com/banzaicloud/bank-vaults/pkg/sdk/vault"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
