+++ operator/pkg/controller/vault/vault_controller.go
@@ -34,8 +34,8 @@ import (
 	"github.com/Masterminds/semver"
 	vaultv1alpha1 "github.com/banzaicloud/bank-vaults/operator/pkg/apis/vault/v1alpha1"
 	"github.com/banzaicloud/bank-vaults/pkg/kv/k8s"
-	bvtls "github.com/banzaicloud/bank-vaults/pkg/tls"
-	"github.com/banzaicloud/bank-vaults/pkg/vault"
+	bvtls "github.com/banzaicloud/bank-vaults/pkg/sdk/tls"
+	"github.com/banzaicloud/bank-vaults/pkg/sdk/vault"
 	"github.com/banzaicloud/k8s-objectmatcher/patch"
 	etcdv1beta2 "github.com/coreos/etcd-operator/pkg/apis/etcd/v1beta2"
 	"github.com/coreos/etcd-operator/pkg/util/etcdutil"
