+++ pkg/kv/kv.go
@@ -21,6 +21,10 @@ type NotFoundError struct {
 	msg string // description of error
 }
 
+func (*NotFoundError) NotFound() bool {
+	return true
+}
+
 func (e *NotFoundError) Error() string { return e.msg }
 
 // NewNotFoundError creates a new NotFoundError
@@ -35,20 +39,4 @@ func NewNotFoundError(msg string, args ...interface{}) *NotFoundError {
 type Service interface {
 	Set(key string, value []byte) error
 	Get(key string) ([]byte, error)
-}
-
-type Tester struct {
-	Service Service
-}
-
-func (t Tester) Test(key string) error {
-	_, err := t.Service.Get(key)
-
-	if err != nil {
-		if _, ok := err.(*NotFoundError); !ok {
-			return err
-		}
-	}
-
-	return t.Service.Set(key, []byte(key))
 }
